extract functions, update to use utils, improve command output
