+++ cli/src/modules/inspect/commonSignatures.ts
@@ -1,0 +1,86 @@
+/**
+ * Standard 4-byte function signatures for common ERC and Diamond interfaces.
+ *
+ * Covers ERC-20, ERC-721, ERC-1155, Diamond Loupe/Cut, and Compose-specific
+ * functions. Used to build the initial selector-to-signature lookup map.
+ */
+export const COMMON_SIGNATURES: readonly string[] = [
+  "acceptOwnership()",
+  "allowance(address,address)",
+  "allowance(address,address,uint256)",
+  "approve(address,uint256)",
+  "approve(address,uint256,uint256)",
+  "balanceOf(address)",
+  "balanceOf(address,uint256)",
+  "balanceOfBatch(address[],uint256[])",
+  "burn(address,uint256)",
+  "burn(address,uint256,uint256)",
+  "burn(uint256)",
+  "burn(uint256,uint256)",
+  "burnBatch(address,uint256[],uint256[])",
+  "burnBatch(uint256[])",
+  "burnFrom(address,uint256)",
+  "burnFrom(address,uint256,uint256)",
+  "checkTokenBridge(address)",
+  "crosschainBurn(address,uint256)",
+  "crosschainMint(address,uint256)",
+  "decimals()",
+  "deleteDefaultRoyalty()",
+  "diamondCut((address,bytes4[],uint8)[],address,bytes)",
+  "DOMAIN_SEPARATOR()",
+  "exportSelectors()",
+  "facetAddress(bytes4)",
+  "facetAddresses()",
+  "facetFunctionSelectors(address)",
+  "facets()",
+  "getApproved(uint256)",
+  "getRoleAdmin(bytes32)",
+  "getRoleExpiry(bytes32,address)",
+  "grantRole(bytes32,address)",
+  "grantRoleBatch(bytes32,address[])",
+  "grantRoleWithExpiry(bytes32,address,uint256)",
+  "hasRole(bytes32,address)",
+  "isApprovedForAll(address,address)",
+  "isOperator(address,address)",
+  "isRoleExpired(bytes32,address)",
+  "isRolePaused(bytes32)",
+  "name()",
+  "nonces(address)",
+  "owner()",
+  "ownerOf(uint256)",
+  "pauseRole(bytes32)",
+  "pendingOwner()",
+  "permit(address,address,uint256,uint256,uint8,bytes32,bytes32)",
+  "renounceOwnership()",
+  "renounceRole(bytes32,address)",
+  "requireRole(bytes32,address)",
+  "requireRoleNotPaused(bytes32,address)",
+  "requireValidRole(bytes32,address)",
+  "resetTokenRoyalty(uint256)",
+  "revokeRole(bytes32,address)",
+  "revokeRoleBatch(bytes32,address[])",
+  "royaltyInfo(uint256,uint256)",
+  "safeBatchTransferFrom(address,address,uint256[],uint256[],bytes)",
+  "safeTransferFrom(address,address,uint256)",
+  "safeTransferFrom(address,address,uint256,bytes)",
+  "safeTransferFrom(address,address,uint256,uint256,bytes)",
+  "setApprovalForAll(address,bool)",
+  "setDefaultRoyalty(address,uint96)",
+  "setOperator(address,bool)",
+  "setRoleAdmin(bytes32,bytes32)",
+  "setTokenRoyalty(uint256,address,uint96)",
+  "supportsInterface(bytes4)",
+  "symbol()",
+  "tokenByIndex(uint256)",
+  "tokenOfOwnerByIndex(address,uint256)",
+  "tokenURI(uint256)",
+  "totalSupply()",
+  "transfer(address,uint256)",
+  "transfer(address,uint256,uint256)",
+  "transferFrom(address,address,uint256)",
+  "transferFrom(address,address,uint256,uint256)",
+  "transferOwnership(address)",
+  "unpauseRole(bytes32)",
+  "upgradeDiamond(address[],(address,address)[],address[],address,bytes,bytes32)",
+  "uri(uint256)",
+];
