+++ cli/src/modules/inspect/module.ts
@@ -1,160 +1,31 @@
-import { isAddress, keccak256, slice, toBytes, type Address, type Hex } from "viem";
 import path from "node:path";
-import { ComposeContext } from "../../context/types";
+import { isAddress, type Address, type Hex } from "viem";
+import { ComposeContext, type ModuleState } from "../../context/types";
 import { DependencyKey } from "../../resolver/dependencyKey";
 import { DependencyResolver } from "../../resolver/dependencyResolver";
 import { resolveChainConfig } from "../../utils/chainConfig";
 import { findFileAncestor } from "../../utils/files";
 import { RPCAdapterError } from "../../adapters/rpc/errors";
 import { showInspect } from "./output";
-import { loadProjectSignatures } from "./abiLoader";
+import { DIAMOND_LOUPE_ABI } from "./diamondLoupeAbi";
+import { toFacetInfo } from "./facetFormatter";
+import { mergeProjectSignatures } from "./selectorDecoder";
 import type { InspectResult, FacetInfo } from "./types";
 
-const DIAMOND_LOUPE_ABI = [
-  {
-    name: "facets",
-    type: "function",
-    stateMutability: "view",
-    inputs: [],
-    outputs: [
-      {
-        type: "tuple[]",
-        components: [
-          { name: "facet", type: "address" },
-          { name: "functionSelectors", type: "bytes4[]" },
-        ],
-      },
-    ],
-  },
-] as const;
-
-const COMMON_SIGNATURES: readonly string[] = [
-  "acceptOwnership()",
-  "allowance(address,address)",
-  "allowance(address,address,uint256)",
-  "approve(address,uint256)",
-  "approve(address,uint256,uint256)",
-  "balanceOf(address)",
-  "balanceOf(address,uint256)",
-  "balanceOfBatch(address[],uint256[])",
-  "burn(address,uint256)",
-  "burn(address,uint256,uint256)",
-  "burn(uint256)",
-  "burn(uint256,uint256)",
-  "burnBatch(address,uint256[],uint256[])",
-  "burnBatch(uint256[])",
-  "burnFrom(address,uint256)",
-  "burnFrom(address,uint256,uint256)",
-  "checkTokenBridge(address)",
-  "crosschainBurn(address,uint256)",
-  "crosschainMint(address,uint256)",
-  "decimals()",
-  "deleteDefaultRoyalty()",
-  "diamondCut((address,bytes4[],uint8)[],address,bytes)",
-  "DOMAIN_SEPARATOR()",
-  "exportSelectors()",
-  "facetAddress(bytes4)",
-  "facetAddresses()",
-  "facetFunctionSelectors(address)",
-  "facets()",
-  "getApproved(uint256)",
-  "getRoleAdmin(bytes32)",
-  "getRoleExpiry(bytes32,address)",
-  "grantRole(bytes32,address)",
-  "grantRoleBatch(bytes32,address[])",
-  "grantRoleWithExpiry(bytes32,address,uint256)",
-  "hasRole(bytes32,address)",
-  "isApprovedForAll(address,address)",
-  "isOperator(address,address)",
-  "isRoleExpired(bytes32,address)",
-  "isRolePaused(bytes32)",
-  "name()",
-  "nonces(address)",
-  "owner()",
-  "ownerOf(uint256)",
-  "pauseRole(bytes32)",
-  "pendingOwner()",
-  "permit(address,address,uint256,uint256,uint8,bytes32,bytes32)",
-  "renounceOwnership()",
-  "renounceRole(bytes32,address)",
-  "requireRole(bytes32,address)",
-  "requireRoleNotPaused(bytes32,address)",
-  "requireValidRole(bytes32,address)",
-  "resetTokenRoyalty(uint256)",
-  "revokeRole(bytes32,address)",
-  "revokeRoleBatch(bytes32,address[])",
-  "royaltyInfo(uint256,uint256)",
-  "safeBatchTransferFrom(address,address,uint256[],uint256[],bytes)",
-  "safeTransferFrom(address,address,uint256)",
-  "safeTransferFrom(address,address,uint256,bytes)",
-  "safeTransferFrom(address,address,uint256,uint256,bytes)",
-  "setApprovalForAll(address,bool)",
-  "setDefaultRoyalty(address,uint96)",
-  "setOperator(address,bool)",
-  "setRoleAdmin(bytes32,bytes32)",
-  "setTokenRoyalty(uint256,address,uint96)",
-  "supportsInterface(bytes4)",
-  "symbol()",
-  "tokenByIndex(uint256)",
-  "tokenOfOwnerByIndex(address,uint256)",
-  "tokenURI(uint256)",
-  "totalSupply()",
-  "transfer(address,uint256)",
-  "transfer(address,uint256,uint256)",
-  "transferFrom(address,address,uint256)",
-  "transferFrom(address,address,uint256,uint256)",
-  "transferOwnership(address)",
-  "unpauseRole(bytes32)",
-  "upgradeDiamond(address[],(address,address)[],address[],address,bytes,bytes32)",
-  "uri(uint256)",
-];
-
-const SELECTOR_MAP = new Map<string, string>();
-
-function ensureSelectorMap(): void {
-  if (SELECTOR_MAP.size > 0) return;
-  for (const sig of COMMON_SIGNATURES) {
-    const trimmed = sig.trim();
-    if (!trimmed) continue;
-    const hash = keccak256(toBytes(trimmed));
-    const selector = slice(hash, 0, 4);
-    SELECTOR_MAP.set(selector.toLowerCase(), trimmed);
-  }
-}
-
-export function decodeSelector(selector: string | Hex): string {
-  ensureSelectorMap();
-  const key = selector.toLowerCase();
-  return SELECTOR_MAP.get(key) ?? selector;
-}
-
-async function mergeProjectSignatures(projectRoot: string): Promise<void> {
-  ensureSelectorMap();
-  const projectMap = await loadProjectSignatures(projectRoot);
-  for (const [selector, signature] of projectMap) {
-    if (!SELECTOR_MAP.has(selector)) {
-      SELECTOR_MAP.set(selector, signature);
-    }
-  }
-}
-
-async function findProjectRoot(): Promise<string | null> {
-  const composePath = await findFileAncestor(process.cwd(), "compose.json");
-  return composePath ? path.dirname(composePath) : null;
-}
-
-function toFacetInfo(raw: { facet: Address; functionSelectors: Hex[] }, index: number): FacetInfo {
-  return {
-    address: raw.facet,
-    index,
-    selectors: raw.functionSelectors.map((sel) => ({
-      selector: sel,
-      signature: decodeSelector(sel),
-    })),
-  };
-}
-
 export const InspectModule = {
+  /**
+   * Inspects an on-chain Diamond and displays its facets and selectors.
+   *
+   * Validates the diamond address, resolves the RPC adapter for the target
+   * chain, fetches facets via the Diamond Loupe, and decodes each selector
+   * using a combination of common signatures and project ABI files.
+   *
+   * @param ctx - The compose context with `address` and optional `chain` params.
+   * @returns The updated context with inspect result stored in
+   *     `ctx.state.inspect` as {@link ModuleState}\<{@link InspectResult}\>.
+   * @throws {RPCAdapterError} If the address is invalid or no contract code is
+   *     found.
+   */
   async inspect(ctx: ComposeContext): Promise<ComposeContext> {
     const addressValue = ctx.param.address;
     if (typeof addressValue !== "string" || !isAddress(addressValue, { strict: false })) {
@@ -193,7 +64,8 @@ export const InspectModule = {
       functionName: "facets",
     });
 
-    const projectRoot = await findProjectRoot();
+    const composePath = await findFileAncestor(process.cwd(), "compose.json");
+    const projectRoot = composePath ? path.dirname(composePath) : null;
     if (projectRoot) {
       await mergeProjectSignatures(projectRoot);
     }
